Basic array functions
Write a program that declare an int array of size 10 inside the main function. Write the
following functions for this array:
1- void initalArry(int arr[]) This function takes an int array parameter. The function will be
used to populate the array by taking input from user.
2- void printArry(int arr[]) Print the array data in an efficient manner that each element is
separated with appropriate space. Note: Use this function to print all results of below
functions.
3- EvenSort(int arr[]). Function to separate even and odd numbers of an array of integers. Put
all even numbers first, and then odd numbers.
4- CheckFrequency(int arr[]) - the function will take an array as input and print the frequency
of each element of an array
5- void binomSum(int arr[],int size) - The function should update the array which contain
numbers equals to the sum of numbers from 1 to the number on each index on array
which is passed as a parameter